Data Object changes for Coverage
