+++ service/build.gradle
@@ -67,6 +67,8 @@ dependencies {
   compile "org.grails.plugins:async"
   compile "org.grails.plugins:events"
   compile "org.grails.plugins:hibernate5"
+  compile 'org.grails.plugins:grails-java8'
+  
   compile "org.hibernate:hibernate-core:5.1.16.Final"
 //  compile "org.hibernate:hibernate-core:5.2.17.Final"
   compile "org.grails.plugins:views-json"
